added new java class
